requirements.txt
beautifulsoup4==4.14.2
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
db-dtypes==1.4.3
et_xmlfile==2.0.0
faiss-cpu==1.12.0
google==3.0.0
google-api-core==2.28.0
google-auth==2.42.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.76.0
idna==3.11
implicit==0.7.2
joblib==1.5.2
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
proto-plus==1.26.1
protobuf==6.33.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
soupsieve==2.8
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
python-dotenv
